# express-main-example/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "0.0.1",
      "title": "Friends List"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": [
      "http",
      "https"
    ],
    "consumes": [
      "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
      "/api/users": {
        "x-swagger-router-controller": "users",
        "get": {
          "description": "Returns a list of paginated users",
          "operationId": "users",
          "parameters": [
            {
              "name": "offset",
              "in": "query",
              "description": "offset is used to skip the number of records from the results.",
              "default": 0,
              "type": "integer"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "offset is used to skip the number of records from the results.",
              "default": 50,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/user"
                },
                "example": [
                  {
                    "id": 0,
                    "firstName": "John",
                    "lastName": "Doe"
                  },
                  {
                    "id": 1,
                    "firstName": "Mark",
                    "lastName": "Zuckerberg"
                  }
                ]
              }
            }
          }
        }
      },
      "/api/friends/{id}": {
        "x-swagger-router-controller": "friends",
        "get": {
          "description": "Returns a list of paginated users",
          "operationId": "friends",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "type": "integer",
              "required": true,
              "description": "Numeric ID of the user to get."
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/user"
                },
                "example": [
                  {
                    "id": 0,
                    "firstName": "John",
                    "lastName": "Doe"
                  },
                  {
                    "id": 1,
                    "firstName": "Mark",
                    "lastName": "Zuckerberg"
                  }
                ]
              }
            },
            "default": {
              "description": "Error",
              "schema": {
                "$ref": "#/definitions/ErrorResponse"
              }
            }
          }
        }
      },
      "/swagger": {
        "x-swagger-pipe": "swagger_raw"
      }
    },
    "definitions": {
      "user": {
        "required": [
          "id",
          "firstName",
          "lastName"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    }
  }